move midi functionality to new file
